refactor: move helpers to folder in dir
