+++ src/helpers/parseFirebaseCollection.ts
