+++ src/helpers/parseFirebaseSnapshot.ts
